Restores proper bundle split for inequality (notably p5)
