+++ src/app/components/elements/modals/inequality/utils.ts
@@ -13,12 +13,11 @@ import {
     EditorMode,
     CHEMICAL_STATES
 } from "./constants";
-import {GREEK_LETTERS_MAP, isDefined, sanitiseInequalityState} from "../../../../services";
+import {GREEK_LETTERS_MAP, isDefined} from "../../../../services";
 import React from "react";
 import isEqual from "lodash/isEqual";
 import uniqWith from "lodash/uniqWith";
-import {Inequality, makeInequality, WidgetSpec} from "inequality";
-import { InequalityState } from "../../inputs/SymbolicTextInput";
+import {Inequality, WidgetSpec} from "inequality";
 
 // This file contains helper functions used specifically in the Inequality modal
 
@@ -680,64 +679,4 @@ export function onCursorMoveEndCallback({movingMenuItem, previousCursor, sketch,
     movingMenuItem.current = null;
     movingMenuBar.current = null;
     potentialSymbolSpec.current = null;
-}
-
-interface PrepareInequalityArgs {
-    editorMode?: EditorMode;
-    logicSyntax?: LogicSyntax;
-    inequalityModalRef: React.RefObject<HTMLDivElement>;
-    isTrashActive: React.MutableRefObject<boolean>;
-    sketch: React.MutableRefObject<Nullable<Inequality>>;
-    initialEditorSymbols: WidgetSpec[];
-    onEditorStateChange?: (state: InequalityState) => void;
-    setEditorState: React.Dispatch<React.SetStateAction<InequalityState>>;
-}
-export function prepareInequality({editorMode, inequalityModalRef, initialEditorSymbols, isTrashActive, sketch, logicSyntax, setEditorState, onEditorStateChange}: PrepareInequalityArgs) {
-    if (!isDefined(inequalityModalRef.current)) {
-        throw new Error("Unable to initialise inequality; target element not found.");
-    }
-    
-    const { sketch: newSketch, p } = makeInequality(
-        inequalityModalRef.current,
-        window.innerWidth,
-        window.innerHeight,
-        initialEditorSymbols,
-        {
-            editorMode: editorMode ?? "logic",
-            logicSyntax: logicSyntax ?? "logic",
-            textEntry: false,
-            fontItalicPath: '/assets/common/fonts/STIXGeneral-Italic.ttf',
-            fontRegularPath: '/assets/common/fonts/STIXGeneral-Regular.ttf'
-        }
-    );
-    if (!isDefined(newSketch)) {
-        throw new Error("Unable to initialize inequality.");
-    }
-    newSketch.log = {
-        initialState: [],
-        actions: [{
-            event: "OPEN",
-            timestamp: Date.now()
-        }]
-    };
-    newSketch.onCloseMenus = () => undefined;
-    newSketch.isTrashActive = () => isTrashActive.current;
-    newSketch.onNewEditorState = (state: InequalityState) => {
-        const modal = inequalityModalRef.current;
-        if (modal) {
-            const newState = sanitiseInequalityState(state);
-            setEditorState((prev: InequalityState) => ({...prev, ...newState}));
-            onEditorStateChange?.(newState);
-        }
-    };
-    sketch.current = newSketch;
-    return () => {
-        if (sketch.current) {
-            sketch.current.onNewEditorState = () => null;
-            sketch.current.onCloseMenus = () => null;
-            sketch.current.isTrashActive = () => false;
-            sketch.current = null;
-        }
-        p.remove();
-    };
 }
